pyproject.toml
[tool.poetry]
name = "qr_query"
version = "0.1.0"
description = ""
authors = ["Ramazan <syrlybai056@gmail.com>"]
readme = "README.md"
packages = [{include = "qr_query"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
django-debug-toolbar = "^4.2.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
